+++ public/class-tmsm-aquatonic-course-booking-public.php
@@ -548,7 +548,7 @@ class Tmsm_Aquatonic_Course_Booking_Public {
 		 * class.
 		 */
 
-		wp_enqueue_style( $this->plugin_name, plugin_dir_url( __FILE__ ) . 'css/tmsm-aquatonic-course-booking-public.css', array(), $this->version, 'all' );
+		wp_enqueue_style( $this->plugin_name, plugin_dir_url( __FILE__ ) . 'css/tmsm-aquatonic-course-booking-public.min.css', array(), $this->version, 'all' );
 
 	}